==> 00_check_setup.ipynb <==
# Check Setup

import sys
import os
import numpy as np
import pandas as pd
import sklearn
import spacy
import nltk
import langdetect

print(f"{sys.version=}")  # Py version of the current environment
print(f"{os.getcwd()=}") ## this should be project's root folder

print(f"{sklearn.__version__=}")
print(f"{nltk.__version__=}")
print(f"{pd.__version__=}")
print(f"{spacy.__version__=}")

# Paths: Data
path_raw = os.path.join("data", "raw")
path_interim = os.path.join("data", "interim")


print(f"{os.path.isdir(path_raw)=}")
print(f"{os.path.isdir(path_interim)=}")

# Input
file_input = "New%20Spanish%20Academic%20Dataset.csv"


path_data_raw = os.path.join(path_raw, file_input)
print(f"{os.path.exists(path_raw)=}")

# Paths: Models
path_models_prod = os.path.join("models", "prod")
path_models_archive = os.path.join("models", "archive")


print(f"{os.path.isdir(path_models_prod)=}")
print(f"{os.path.isdir(path_models_archive)=}")
