==> 01_load_and_split.ipynb <==
# Load and split data

import os
import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split


# Parameters

RND_SEED = 123
PCT_TEST = 0.2
TARGET_MAP = {"F": 0, "NF": 1}
# Paths
path_raw = os.path.join("data", "raw")
path_interim = os.path.join("data", "interim")

# Input
file_input = "New%20Spanish%20Academic%20Dataset.csv"

# Output
file_prep = "prep.csv"
file_train = "train.csv"
file_test = "test.csv"

# Load data

path_data_raw = os.path.join(path_raw, file_input)

df_raw = pd.read_csv(...)
df_raw.head()

# Obtain the number or rows and columns


# Calculate the percentage of N and NF rows


# Transform data

df_prep = df_raw.copy()
df_prep['y_is_nf'] = ...
df_prep['x_text'] = ...
df_prep = df_prep[["x_text", "y_is_nf"]]

# Split data

# add parametrized test_size and random_state and add a propper sampling stratgy
# when the input is a single object, the output will maintain the same structure
df_train, df_test = train_test_split(
    df_prep,
    ...
)

print(f"{df_train.shape=}")
print(f"{df_test.shape=}")

# Write Data

if not os.path.exists(path_interim):
    print("Please, create the folder /data/interim/ !")

df_prep.to_csv(
    os.path.join(path_interim, file_prep),
    index=False)

df_train.to_csv(
    os.path.join(path_interim, file_train),
    index=False)

df_test.to_csv(
    os.path.join(path_interim, file_test),
    index=False)
